requirements.txt
anthropic>=0.40.0
orq-ai-sdk>=3.0.0
chromadb>=0.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0
